refactor: components in dedicated directory
