+++ pages/impactlivraison.js
@@ -1,16 +1,18 @@
 import categories from 'data/categories.json'
 
 import Section2 from 'components/base/Section2'
-
-import CalculateurLivraison from './impactlivraison/CalculateurLivraison'
-import IntroLivraison from './impactlivraison/IntroLivraison'
+import CalculateurLivraison from 'components/livraison/CalculateurLivraison'
+import IntroLivraison from 'components/livraison/IntroLivraison'
+import { RulesProvider } from 'components/livraison/RulesProvider'
 
 export default function Impactlivraison() {
   return (
     <Section2>
       <Section2.InnerMargin>
-        <IntroLivraison />
-        <CalculateurLivraison />
+        <RulesProvider>
+          <IntroLivraison />
+          <CalculateurLivraison />
+        </RulesProvider>
       </Section2.InnerMargin>
     </Section2>
   )
